requirements.txt
aiohttp
configparser

# database packages
networkx
aiomysql

# Basic modules
numpy
pipmaster
pydantic
python-dotenv

setuptools
tenacity

# LLM packages
tiktoken

# Extra libraries are installed when needed using pipmaster
